Create auth rules for controller actions

+++ controllers/bird_controller.py
@@ -1,9 +1,11 @@
 from flask import Blueprint, request
 from flask_jwt_extended import get_jwt_identity, jwt_required
-from functions import find_entity_by_id
+from functions import delete_admin_entity, find_all_entities, find_entity_by_id
 
 from init import db
+from models.approved_bird import ApprovedBird
 from models.bird import Bird, bird_schema, birds_schema
+from models.user import User
 
 
 birds_bp = Blueprint("birds", __name__, url_prefix="/birds")
@@ -11,8 +13,7 @@ birds_bp = Blueprint("birds", __name__, url_prefix="/birds")
 
 @birds_bp.route("/", methods=["GET"])
 def get_all_birds():
-    stmt = db.select(Bird).order_by(Bird.id)
-    birds = db.session.scalars(stmt)
+    birds = find_all_entities(Bird, Bird.id)
     return birds_schema.dump(birds)
 
 
@@ -41,12 +42,21 @@ def create_bird():
 
 
 @birds_bp.route("/<int:id>", methods=["DELETE"])
+@jwt_required()
 def delete_bird(id):
     bird = find_entity_by_id(Bird, id)
     if bird:
+        # Only admins allowed to delete approved birds
+        if bird.is_approved == True:
+            return delete_admin_entity(bird)
+        # Only allows original user or admin to delete birds
+        user = find_entity_by_id(User, get_jwt_identity())
+        if user.id != bird.submitting_user_id and not user.is_admin:
+            return {"error": "Not authorised to perform action. Only original user may delete."}, 403
+        
         db.session.delete(bird)
         db.session.commit()
-        return {"message": f"Bird '{bird.name}' deleted successfully!"}
+        return {"message":f"Bird with id: {id} successfully deleted"}
     else:
         return {"error": f"Bird with id: {id} not found"}
 
@@ -54,16 +64,21 @@ def delete_bird(id):
 @birds_bp.route("/<int:id>", methods=["PUT", "PATCH"])
 @jwt_required()
 def update_bird(id):
+    # gets body data
     body_data = bird_schema.load(request.get_json(), partial=True)
-
+    # finds bird in table
     bird = find_entity_by_id(Bird, id)
-    if not bird:
-        return {"error": f"Bird with id: {id} not found"}
-
+    #  updates details, toggles is_approved
     bird.name = body_data.get("name") or bird.name
     bird.description = body_data.get("description") or bird.description
     bird.submitting_user_id = get_jwt_identity()
     bird.is_approved = False
+
+    # removes bird from ApprovedBird table
+    stmt = db.select(ApprovedBird).filter_by(bird_id=id)
+    approved_bird = db.session.scalar(stmt)
+    if approved_bird:
+        db.session.delete(approved_bird)
 
     db.session.commit()
     return bird_schema.dump(bird)
